Remove the /plans/<plan_id> prefix again
Some validations (e.g. initial hardware checks) may not be associated
with a plan so requiring it unconditionally is not a good idea.

This removes the `/plans/` route, but it lets the `run` and `stop`
commands accept a `plan_id` URL parameter. Its value is then reflected
in the result (to clarify which plan the result belongs to).