Add the rogue DHCP validation
This adds a new validation that checks for DHCP responses on the
undercloud. Having a DHCP on a Pacemaker-managed network is a known
incompatibility.

Eventually, we will want to get a list of all the networks used for the
deployment and check for DHCP conflicts. In the meantime, this just
returns a list of discovered DHCP servers back.