Temporarily disable the Discovery Diff validation
I can't seem to get it running right now and it seems to prevent all the
other validations from the "hardware-discovery" stage as well, so we're
disabling it for a moment.

It doesn't seem to be just a failing validation either -- that would let
the other ones run -- but something more severe.