Check for the MySQL open files limit
Based on:

https://access.redhat.com/solutions/1598733

"The default file descriptor limit for MariaDB/Galera is too low for an
OpenStack deployment. When the database server hits this limit, it
cannot open new file descriptors to process additional database
connections and will refuse them from various OpenStack services causing
failure of the associated OpenStack tasks. It is recommended that you
increase this limit to 16384."