Add Ansible output to the API
**Candidate for an ansible1.9 backport.**

This uses a custom Runner callback to capture the messages (including
warnings, failure reasons, etc.) Ansible normally prints to stdout.

These messages are then slightly filtered, processed and passed to the
API. The `result` API object contains 3 more objects from now on: `raw`
which contains the messages unprocessed, `failure_messages` which is a
list of messages from task failures and `warning_messages` which
collects all the warnings the tasks produced.

This will let the callers show more useful information on the failed
validations (e.g. the actual reason for the failure).