Install scapy into a virtualenv
Some systems (e.g. RHEL 7) don't ship pip, which means the Rogue DHCP
validation could not install its dependencies. This changes the validation to
create a virtualenv and install scapy there.

Unfortunately, I couldn't find a way to have Ansible modules include this new
virtualenv, so this change includes rewriting rogue_dhcp.py to be a standalone
script executed within the virtualenv using the `command` Ansible module
instead of treating it as a normal module.

That means we lose some of the finer-grained reporting capabilities. In
addition, it looks like our PoC API (validation-api.py) doesn't report STDERR
or anything useful in situations like this -- that will be addressed in the
follow-up commits.