Return results for more than one host
When there is more than one host for the validation, Ansible forks the process
that invoked it that many times and this makes any changes we make to a shared
data structure disappear. Which results in the API not returning any details
about the resulting runs.

By setting `forks=1` we prevent this and everything is run from a single
process. This would not be advisable in production but should be fine as a
proof of concept.