Add additional files to neutron-sanity-check
neutron-sanity-check expects to be passed all the --config-file/--config-dir
options the Neutron services receive.

Unfortunately, when we pass in a file/directory that doesn't exist, the script
crashes. So we're building the list of existing files/dirs in a shell script
and pass only the ones they're present.

We could do the same in Ansible, but that would generate a lot of duplicate
output and be more confusing to the users.