Fix the Pacemaker validation under Ansible 1.9
There's a bug in Ansible 1.9 where passing some XML content between
tasks or modules like we do here breaks. Running this validation could
result in failing with this error:

    a duplicate parameter was found in the argument string (2016" exec-time)

The error happened while running the "pacemaker" custom module, but was
present even when I removed everything but the module setup code.

This seems to be fixed in Ansible 2.0, but for 1.9, encoding the
pacemaker output in base64 fixes the problem.